+++ src/cli/expressions/InterpreterExpression.js
@@ -1,0 +1,5 @@
+module.exports = class InterpreterExpression {
+  interpret(context /* eslint-disable-line no-unused-vars */) {
+    throw new Error("Not Implemented");
+  }
+};
